refactor(mcp): fold mcp-config into mcp-contract
